Clean Up
Created 2 extra clalles for extenstion methid for Enemy_AI,
and one to store DataHolder since it was getting big.
More improvements to AI list sorting.
