+++ Dots-RTS-Development/Assets/BulkRename.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 15eedef553a4a5c4c9ac76d8104e5807
+folderAsset: yes
+timeCreated: 1506764672
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
