+++ Dots-RTS-Development/Assets/Scripts/AI/AI_Data_Holder.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 30a93200b4f0b7543bba97fe92a8542e
+timeCreated: 1507237005
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
